requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.6
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
Django==5.1.1
django-stubs==5.1.0
django-stubs-ext==5.1.0
djangorestframework==3.15.2
exceptiongroup==1.2.2
fastapi==0.115.0
frozenlist==1.4.1
h11==0.14.0
idna==3.10
multidict==6.1.0
pydantic==2.9.2
pydantic_core==2.23.4
requests==2.32.3
sniffio==1.3.1
sqlparse==0.5.1
starlette==0.38.6
tomli==2.0.2
types-PyYAML==6.0.12.20240917
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.31.0
yarl==1.12.1
